requirements.txt
absl-py==0.7.1
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
audioread==2.1.8
beautifulsoup4==4.8.0
cffi==1.12.3
clldutils==3.1.1
colorlog==4.0.2
csvw==1.6.0
cycler==0.10.0
decorator==4.4.0
fastdtw==0.3.4
future==0.18.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.9.0
importlib-metadata==0.20
isodate==0.6.0
joblib==0.13.2
kiwisolver==1.1.0
librosa==0.7.0
llvmlite==0.29.0
Markdown==3.1.1
matplotlib==3.1.1
more-itertools==7.2.0
numba==0.45.1
numpy==1.16.4
packaging==19.1
pandas==1.0.1
Pillow==6.1.0
pkg-resources==0.0.0
playsound==1.2.2
pluggy==0.12.0
protobuf==3.9.1
py==1.8.0
pycparser==2.19
pydeps==1.7.3
pyparsing==2.4.2
pytest==5.1.2
python-dateutil==2.8.0
pytorch-ignite==0.3.0
pytorch-lightning==0.6.0
pytz==2019.3
PyYAML==5.3
regex==2019.8.19
resampy==0.2.2
rfc3986==1.3.2
scikit-learn==0.21.3
scipy==1.3.1
segments==2.1.0
six==1.12.0
sklearn==0.0
SoundFile==0.10.2
soupsieve==1.9.3
stdlib-list==0.5.0
tabulate==0.8.5
tb-nightly==1.14.0a20190603
tensorboard==2.0.0
termcolor==1.1.0
torch==1.2.0
torch-stft==0.1.4
torchaudio==0.3.0
torchcontrib==0.0.2
torchnmf==0.2
torchvision==0.4.0
tqdm==4.42.1
uritemplate==3.0.0
wcwidth==0.1.7
Werkzeug==0.15.5
wrapt==1.11.2
zipp==0.6.0
barbar==0.2.1
gitdb==4.0.2
g2p-en==2.1.0
